feat: extract bin code for kafka create topic and pg create and migrate into reusable utils
