+++ packages/talos_certifier_adapters/Cargo.toml
@@ -31,6 +31,7 @@ tokio-postgres = { version = "0.7", features = [
     "with-serde_json-1",
 ] }
 deadpool-postgres = "0.10"
+refinery = { version = "0.8.7", features = ["tokio-postgres"] }
 
 # Error
 thiserror = "1.0.31"
