+++ packages/talos_certifier_adapters/src/kafka/mod.rs
@@ -1,5 +1,6 @@
 pub mod config;
 pub mod consumer;
 pub mod errors;
+pub mod kafka_deploy;
 pub mod producer;
 pub mod utils;
