+++ packages/talos_certifier_adapters/src/lib.rs
@@ -1,12 +1,12 @@
 // Kafka exports
-mod kafka;
+pub mod kafka;
 pub use kafka::config::Config as KakfaConfig;
 pub use kafka::consumer::KafkaConsumer;
 pub use kafka::errors::KafkaAdapterError;
 pub use kafka::producer::KafkaProducer;
 
 // postgres exports
-mod postgres;
+pub mod postgres;
 pub use postgres::config::PgConfig;
 pub use postgres::errors::PgError;
 pub use postgres::pg::Pg;
