+++ packages/talos_certifier_adapters/src/postgres/mod.rs
@@ -1,4 +1,7 @@
 pub mod config;
 pub mod errors;
 pub mod pg;
+mod pg_deploy;
 mod utils;
+
+pub use pg_deploy::{create_database, run_migration};
